+++ docs/changelog.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ## What's Changed
 * kw: send decodes what it is about to broadcast
+* sync: a header must carry the difficulty the chain demands of it
 
 ## [0.2.3] - 2026-09-11
 ## What's Changed